Corrigir formulário de destaque de imóvel e restaurar página meus-imoveis

+++ wp-content/themes/socasatop/inc/custom/broker/shortcodes.php
@@ -249,38 +249,30 @@ function highlight_payment_form_shortcode($atts) {
     }
     
     // Incluir os scripts e estilos necessários
-    wp_enqueue_style('highlight-css', get_stylesheet_directory_uri() . '/inc/custom/broker/assets/css/highlight.css', array(), '1.0.5');
+    wp_enqueue_style('highlight-css', get_stylesheet_directory_uri() . '/inc/custom/broker/assets/css/highlight.css', array(), '1.0.7');
     wp_enqueue_script('mercadopago-js', 'https://sdk.mercadopago.com/js/v2', array(), null, true);
-    wp_enqueue_script('highlight-payment-js', get_stylesheet_directory_uri() . '/inc/custom/broker/assets/js/highlight-payment.js', array('jquery'), '1.0.5', true);
+    wp_enqueue_script('highlight-payment-js', get_stylesheet_directory_uri() . '/inc/custom/broker/assets/js/highlight-payment.js', array('jquery'), '1.0.7', true);
     
-    // Passar dados para o JavaScript
-    $highlight_price = get_option('highlight_payment_price', 149.90);
-    $mp_public_key = get_option('mercadopago_public_key_test');
-    
-    if (get_option('mercadopago_production_mode', 'no') === 'yes') {
-        $mp_public_key = get_option('mercadopago_public_key_prod');
+    // Verificar se o arquivo de processamento do destaque existe
+    $highlight_file = get_template_directory() . '/inc/custom/broker/highlight-payment.php';
+    if (!file_exists($highlight_file)) {
+        error_log('Arquivo de destaque não encontrado: ' . $highlight_file);
+        return '<div class="error-message">Sistema de destaque temporariamente indisponível.</div>';
     }
     
-    wp_localize_script('highlight-payment-js', 'highlight_payment', array(
-        'ajax_url' => admin_url('admin-ajax.php'),
-        'nonce' => wp_create_nonce('highlight_payment_nonce'),
-        'immobile_id' => $immobile_id,
-        'price' => $highlight_price,
-        'public_key' => $mp_public_key,
-    ));
+    // Incluir o arquivo do sistema de destaque
+    require_once($highlight_file);
     
     // Usar diretamente a função de renderização do formulário
     ob_start();
-    echo '<div class="highlight-payment-wrapper">';
     
     if (function_exists('render_highlight_payment_form')) {
-        echo render_highlight_payment_form(array('immobile_id' => $immobile_id));
+        echo render_highlight_payment_form($immobile_id);
     } else {
         echo '<div class="error-message">O sistema de pagamento para destaque está temporariamente indisponível.</div>';
         error_log('Função render_highlight_payment_form não encontrada');
     }
     
-    echo '</div>';
     return ob_get_clean();
 }
 add_shortcode('highlight_payment_form', 'highlight_payment_form_shortcode');
